Math Formula: Split add-on in multiple files
