+++ math_formula/main.py
@@ -1,20 +1,7 @@
 import bpy
 import blf
-import bgl
-import math
-import rna_keymap_ui
 
-
-bl_info = {
-    "name": "Node Math Formula",
-    "author": "Wannes Malfait",
-    "version": (0, 5, 0),
-    "location": "Node Editor Toolbar",
-    "description": "Quickly add math nodes by typing in a formula",
-    "category": "Node",
-    "blender": (2, 93, 0),  # Required so the add-on will actually load
-}
-
+from .positioning import MFPositionNode, MFTreePositioner
 
 # Operations used by the math node
 math_operations = [
@@ -100,131 +87,6 @@ builtin_attributes = [
 ]
 
 formula_history = []
-
-
-class MFMathFormula(bpy.types.AddonPreferences):
-    bl_idname = __name__
-
-    font_size: bpy.props.IntProperty(
-        name="Font Size",
-        description="Font size used for displaying text",
-        default=15,
-        min=8,
-        soft_max=20,
-    )
-    node_distance: bpy.props.IntProperty(
-        name="Distance between nodes",
-        description="The distance placed between the nodes from the formula",
-        default=30,
-        min=0,
-    )
-    sibling_distance: bpy.props.IntProperty(
-        name="Distance between siblings",
-        description="The distance between nodes which connect to the same node",
-        default=20,
-        min=0,
-    )
-    subtree_distance: bpy.props.IntProperty(
-        name="Distance between subtrees",
-        description="The distance between two subtrees",
-        default=50,
-        min=0,
-    )
-    show_colors: bpy.props.BoolProperty(
-        name="Show colors for syntax highlighting",
-        default=False,
-    )
-    builtin_attr_color: bpy.props.FloatVectorProperty(
-        name="Built-in Attribute Color",
-        default=(0.5, 0.3, 0.05),
-        subtype='COLOR',
-    )
-    math_func_color: bpy.props.FloatVectorProperty(
-        name="Math Function Color",
-        default=(0.0, 0.8, 0.1),
-        subtype='COLOR',
-    )
-    vector_math_func_color: bpy.props.FloatVectorProperty(
-        name="Vector Math Function Color",
-        default=(0.142, 0.408, 0.8),
-        subtype='COLOR',
-    )
-    grouping_color: bpy.props.FloatVectorProperty(
-        name="Grouping Indicator Color",
-        default=(0.3, 0.1, 0.8),
-        subtype='COLOR',
-    )
-    separate_combine_color: bpy.props.FloatVectorProperty(
-        name="Combine Separate XYZ Color",
-        default=(0.76, 0.195, 0.071),
-        subtype='COLOR',
-    )
-    float_color: bpy.props.FloatVectorProperty(
-        name="Number Color",
-        default=(0.7, 0.515, 0.462),
-        subtype='COLOR',
-    )
-    default_color: bpy.props.FloatVectorProperty(
-        name="Default Color",
-        default=(1.0, 1.0, 1.0),
-        subtype='COLOR',
-    )
-    result_color: bpy.props.FloatVectorProperty(
-        name="Result Color",
-        default=(0.103, 0.8, 0.492),
-        subtype='COLOR',
-    )
-
-    def draw(self, context):
-        layout = self.layout
-        col = layout.column()
-        col.prop(self, 'font_size')
-        col.prop(self, 'node_distance')
-        col.prop(self, 'sibling_distance')
-        col.prop(self, 'subtree_distance')
-        col.separator()
-        col.prop(self, 'show_colors')
-        if self.show_colors:
-            box = layout.box()
-            box.label(text="Syntax Highlighting")
-            box.prop(self, 'builtin_attr_color')
-            box.prop(self, 'math_func_color')
-            box.prop(self, 'vector_math_func_color')
-            box.prop(self, 'grouping_color')
-            box.prop(self, 'separate_combine_color')
-            box.prop(self, 'float_color')
-            box.prop(self, 'default_color')
-            box.prop(self, 'result_color')
-        col = layout.column()
-        col.label(text="Keymaps:")
-        kc = bpy.context.window_manager.keyconfigs.addon
-        for km, kmi in addon_keymaps:
-            km = km.active()
-            col.context_pointer_set("keymap", km)
-            rna_keymap_ui.draw_kmi([], kc, km, kmi, col, 0)
-
-
-class MF_Settings(bpy.types.PropertyGroup):
-    formula: bpy.props.StringProperty(
-        name="Formula",
-        description="Formula written in Reverse Polish Notation",
-        default="4 5 *",
-    )
-    temp_attr_name: bpy.props.StringProperty(
-        name="Temporary Attribute",
-        description="Name of the temporary attribute used to store in between results",
-        default="tmp",
-    )
-    no_arg: bpy.props.StringProperty(
-        name="Missing Argument Name",
-        default="",
-        description="The name of the attribute used to fill in missing arguments"
-    )
-    add_frame: bpy.props.BoolProperty(
-        name="Add Frame",
-        description='Put all the nodes in a frame',
-        default=False,
-    )
 
 
 def mf_check(context) -> bool:
@@ -469,260 +331,6 @@ class MFParser:
         Try updating the last token with the new token from the string
         """
         pass
-
-
-class MFPositionNode():
-    def __init__(self, node: bpy.types.Node, parent=None, children=None, left_sibling=None, right_sibling=None, has_dimensions=False):
-        self.node = node
-        self.parent: MFPositionNode = parent
-        self.children: list[MFPositionNode] = children
-        self.first_child: MFPositionNode = children[0] if children else None
-        self.left_sibling: MFPositionNode = left_sibling
-        self.right_sibling: MFPositionNode = right_sibling
-        # TODO: make update work so this is correct and doesn't require such a hack
-        if has_dimensions:
-            self.width = self.node.dimensions.x
-            self.height = self.node.dimensions.y
-        else:
-            inputs = 0
-            linked_sockets = 0
-            for socket in node.inputs:
-                if socket.enabled:
-                    inputs += 1
-                if socket.is_linked:
-                    linked_sockets += 1
-            if 'NodeMath' in node.bl_idname:
-                self.height = 120
-                self.height += inputs*22
-            elif 'NodeVectorMath' in node.bl_idname:
-                self.height = 96
-                self.height += inputs*88
-                self.height -= linked_sockets*66
-            self.width = 153.611
-        self.prelim_y = 0
-        self.modifier = 0
-        self.left_neighbour: MFPositionNode = None
-
-    def set_x(self, x):
-        self.node.location.x = x
-
-    def set_y(self, y):
-        self.node.location.y = y
-
-    def get_x(self) -> int:
-        return self.node.location.x
-
-    def get_y(self) -> int:
-        return self.node.location.y
-
-    def get_width(self) -> float:
-        return self.width
-
-    def get_height(self) -> float:
-        return self.height
-
-    def is_leaf(self) -> bool:
-        return self.first_child is None
-
-    def has_right(self) -> bool:
-        return self.right_sibling is not None
-
-    def has_left(self) -> bool:
-        return self.left_sibling is not None
-
-    def __str__(self) -> str:
-        parent = self.parent.node.operation if self.parent else ""
-        # first_child = self.first_child.node.operation if self.first_child else ""
-        left = self.left_sibling.node.operation if self.left_sibling else ""
-        right = self.right_sibling.node.operation if self.right_sibling else ""
-        neighbour = self.left_neighbour.node.operation if self.left_neighbour else ""
-        return f"{self.node.operation}:\n \
-        parent: {parent}, \n \
-        children: {self.children}, \n \
-        left sibling: {left}, \n \
-        right sibling: {right}, \n \
-        left neighbour: {neighbour}"
-
-    def __repr__(self) -> str:
-        return f"{self.node.operation}"
-
-
-class MFTreePositioner():
-    """
-    Class to position nodes in a node tree
-    Algorithm: https://www.cs.unc.edu/techreports/89-034.pdf
-    """
-
-    def __init__(self, context):
-        prefs = context.preferences.addons[__name__].preferences
-        self.level_separation: int = prefs.node_distance
-        self.sibling_separation: int = prefs.sibling_distance
-        self.subtree_separation: int = prefs.subtree_distance
-        self.x_top_adjustment: int = 0
-        self.y_top_adjustment: int = 0
-        self.max_width_per_level: list[float] = [0 for _ in range(100)]
-        self.prev_node_per_level = [None for _ in range(100)]
-        self.min_x_loc = +math.inf
-        self.max_x_loc = -math.inf
-        self.min_y_loc = +math.inf
-        self.max_y_loc = -math.inf
-        self.visited_nodes: list[MFPositionNode] = []
-
-    def place_nodes(self, root_node: MFPositionNode, cursor_loc: tuple[float] = None) -> tuple[float]:
-        """
-        Aranges the nodes connected to `root_node` so that the top
-        left corner lines up with `cursor_loc`. If `cursor_loc` is `None`,
-        the tree is aligned such that `root_node` stays in the same place.
-
-        The returned value is the top right corner, i.e the place where
-        you would want to place the next nodes, if `cursor_loc` is not `None`.
-        Otherwise `None` is returned.
-        """
-        old_root_node_pos_x, old_root_node_pos_y = root_node.node.location
-        self.first_walk(root_node, 0)
-        self.x_top_adjustment = root_node.get_x()
-        self.y_top_adjustment = root_node.get_y() - root_node.prelim_y
-        self.second_walk(root_node, 0, 0, 0)
-        offset_x = 0
-        offset_y = 0
-        if cursor_loc is not None:
-            offset_x = cursor_loc[0] - self.min_x_loc
-            offset_y = cursor_loc[1] - self.max_y_loc
-        else:
-            offset_x = old_root_node_pos_x-root_node.get_x()
-            offset_y = old_root_node_pos_y-root_node.get_y()
-        for pnode in self.visited_nodes:
-            pnode.set_x(pnode.get_x() + offset_x)
-            pnode.set_y(pnode.get_y() + offset_y)
-        if cursor_loc is not None:
-            return (cursor_loc[0]+self.max_x_loc-self.min_x_loc, cursor_loc[1])
-
-    def get_leftmost(self, node: MFPositionNode, level: int, depth: int) -> MFPositionNode:
-        if level >= depth:
-            return node
-        if node.is_leaf():
-            return None
-        rightmost = node.first_child
-        leftmost = self.get_leftmost(rightmost, level + 1, depth)
-        while leftmost is None and rightmost.has_right():
-            rightmost = rightmost.right_sibling
-            leftmost = self.get_leftmost(rightmost, level + 1, depth)
-        return leftmost
-
-    def get_prev_node_at_level(self, level: int) -> MFPositionNode:
-        return self.prev_node_per_level[level]
-
-    def set_prev_node_at_level(self, level: int, node: MFPositionNode):
-        self.prev_node_per_level[level] = node
-
-    def apportion(self, node: MFPositionNode):
-        leftmost = node.first_child
-        neighbour = leftmost.left_neighbour
-        compare_depth = 1
-        while leftmost is not None and neighbour is not None:
-            # Compute the location of leftmost and where it
-            # should be with respect to neighbour
-            left_mod_sum = right_mod_sum = 0
-            ancestor_leftmost = leftmost
-            ancestor_neighbour = neighbour
-
-            for _ in range(compare_depth):
-                ancestor_leftmost = ancestor_leftmost.parent
-                ancestor_neighbour = ancestor_neighbour.parent
-                right_mod_sum += ancestor_leftmost.modifier
-
-                left_mod_sum += ancestor_neighbour.modifier
-
-            # Find the move_distance and apply it to the node's subtree
-            # Add appropriate portions to smaller interior subtrees
-            move_distance = neighbour.prelim_y +\
-                left_mod_sum + \
-                self.subtree_separation + \
-                neighbour.get_height() - \
-                (leftmost.prelim_y + right_mod_sum)
-            if move_distance > 0:
-                tmp = node
-                left_siblings = 0
-                # Count the interior sibling subtrees
-                while tmp is not None and tmp != ancestor_neighbour:
-                    left_siblings += 1
-                    tmp = tmp.left_sibling
-                if tmp is not None:
-                    # Apply posrtions to appropriate left sibling
-                    # subtrees
-                    portion = move_distance/left_siblings
-                    tmp = node
-                    while tmp != ancestor_neighbour:
-                        tmp.prelim_y += move_distance
-                        tmp.modifier += move_distance
-                        move_distance -= portion
-                        tmp = tmp.left_sibling
-                else:
-                    # In this case ancestor_neighbour and ancestor_leftmost
-                    # aren't siblings, so the job to move should be done by
-                    # an ancestor instead
-                    return
-
-            # Determine the leftmost descendant of Node at the next lower level
-            # to compare its positioning against that of its neighbour.
-            compare_depth += 1
-            if leftmost.is_leaf():
-                leftmost = self.get_leftmost(node, 0, compare_depth)
-            else:
-                leftmost = leftmost.first_child
-            if leftmost is not None:
-                neighbour = leftmost.left_neighbour
-            else:
-                return
-
-    def first_walk(self, node: MFPositionNode, level: int):
-        node.left_neighbour = self.get_prev_node_at_level(level)
-        self.set_prev_node_at_level(level, node)
-        node.modifier = 0
-        if node.is_leaf():
-            if node.has_left():
-                node.prelim_y = node.left_sibling.prelim_y + \
-                    self.sibling_separation + \
-                    node.left_sibling.get_height()
-            else:
-                node.prelim_y = 0
-        else:
-            # It's not a leaf, so recursivly call for children
-            leftmost = rightmost = node.first_child
-            self.first_walk(leftmost, level+1)
-            while rightmost.has_right():
-                rightmost = rightmost.right_sibling
-                self.first_walk(rightmost, level+1)
-            mid = (leftmost.prelim_y + rightmost.prelim_y)/2
-            if node.has_left():
-                node.prelim_y = node.left_sibling.prelim_y + \
-                    self.sibling_separation + \
-                    node.left_sibling.get_height()
-                node.modifier = node.prelim_y - mid
-                self.apportion(node)
-            else:
-                node.prelim_y = mid
-        self.max_width_per_level[level] = max(
-            node.width, self.max_width_per_level[level])
-
-    def second_walk(self, node: MFPositionNode, level: int, width_sum_x: float, mod_sum_y: float):
-        x = self.x_top_adjustment - width_sum_x
-        y = self.y_top_adjustment - node.prelim_y - mod_sum_y
-        self.min_x_loc = min(x, self.min_x_loc)
-        self.min_y_loc = min(y+node.get_height(), self.min_y_loc)
-        self.max_x_loc = max(x+node.get_width(), self.max_x_loc)
-        self.max_y_loc = max(y, self.max_y_loc)
-        node.set_x(x)
-        node.set_y(y)
-        self.visited_nodes.append(node)
-        if not node.is_leaf():
-            self.second_walk(node.first_child, level + 1,
-                             width_sum_x +
-                             self.max_width_per_level[level+1] +
-                             self.level_separation,
-                             mod_sum_y + node.modifier)
-        if node.has_right():
-            self.second_walk(node.right_sibling, level, width_sum_x, mod_sum_y)
 
 
 class MF_OT_arrange_from_root(bpy.types.Operator, MFBase):
@@ -1233,36 +841,6 @@ class MF_OT_attribute_math_formula_add(bpy.types.Operator, MFBase):
         return self.execute(context)
 
 
-class MF_PT_panel(bpy.types.Panel, MFBase):
-    bl_idname = "NODE_PT_mf_math_formula"
-    bl_space_type = 'NODE_EDITOR'
-    bl_label = "Add Math Formula"
-    bl_region_type = "UI"
-    bl_category = "Math Formula"
-
-    def draw(self, context):
-
-        # Helper variables
-        layout = self.layout
-        scene = context.scene
-        props = scene.math_formula_add
-
-        col = layout.column(align=True)
-        col.label(text="Addon Preferences has more settings")
-        col.prop(props, 'formula')
-        col.prop(props, 'add_frame')
-        col.separator()
-        if context.space_data.tree_type == 'GeometryNodeTree':
-            col.prop(props, 'temp_attr_name')
-            col.prop(props, 'no_arg')
-            col.operator(MF_OT_attribute_math_formula_add.bl_idname)
-        col.operator(MF_OT_math_formula_add.bl_idname)
-        if context.active_node is not None:
-            col.operator(MF_OT_arrange_from_root.bl_idname)
-        else:
-            col.label(text="No active node")
-
-
 def draw_callback_px(self,):
     font_id = 0
     font_size = self.font_size
@@ -1456,24 +1034,10 @@ class MF_OT_type_formula_then_add_nodes(bpy.types.Operator, MFBase):
         return {'RUNNING_MODAL'}
 
 
-addon_keymaps = []
-kmi_defs = [
-    # kmi_defs entry: (identifier, key, action, CTRL, SHIFT, ALT, props)
-    # props entry: (property name, property value)
-    (MF_OT_arrange_from_root.bl_idname, 'E', 'PRESS', False, False, True, None),
-    (MF_OT_type_formula_then_add_nodes.bl_idname,
-     'F', 'PRESS', False, True, False, (('use_attributes', True),)),
-    (MF_OT_type_formula_then_add_nodes.bl_idname,
-     'F', 'PRESS', False, False, True, (('use_attributes', False),))
-]
-
 classes = (
-    MFMathFormula,
-    MF_Settings,
     MF_OT_arrange_from_root,
     MF_OT_attribute_math_formula_add,
     MF_OT_math_formula_add,
-    MF_PT_panel,
     MF_OT_type_formula_then_add_nodes,
 )
 
@@ -1482,34 +1046,7 @@ def register():
     for cls in classes:
         bpy.utils.register_class(cls)
 
-    # keymaps (Code from node wrangler)
-    addon_keymaps.clear()
-    kc = bpy.context.window_manager.keyconfigs.addon
-    if kc:
-        km = kc.keymaps.new(name='Node Editor', space_type="NODE_EDITOR")
-        for (identifier, key, action, CTRL, SHIFT, ALT, props) in kmi_defs:
-            kmi = km.keymap_items.new(
-                identifier, key, action, ctrl=CTRL, shift=SHIFT, alt=ALT)
-            kmi.active = True
-            if props:
-                for prop, value in props:
-                    setattr(kmi.properties, prop, value)
-            addon_keymaps.append((km, kmi))
-
-    bpy.types.Scene.math_formula_add = bpy.props.PointerProperty(
-        type=MF_Settings)
-
 
 def unregister():
     for cls in classes:
         bpy.utils.unregister_class(cls)
-
-    for km, kmi in addon_keymaps:
-        km.keymap_items.remove(kmi)
-    addon_keymaps.clear()
-
-    del bpy.types.Scene.math_formula_add
-
-
-if __name__ == "__main__":
-    register()
